Create Controls.java
This version does not produce a result. It assumes beginning value of 0. It allows values to be entered through the app. Text or comments, the lower textbox may be entered only through the keyboard. The equals does not work at this time. It may be changed to result, where it parses the arraylist and then recalculates the result. The clear button will clear the textboxes, but also cause an error to be thrown on when the next operand is pressed.