2014.11.06@10:02PM Update
This update utilizes the file towel-v1.2.2.jar found here: https://github.com/MarkyVasconcelos/Towel/downloads. This allows for the conversion of the string equation to an actionable calculation. This version allows one complete equation and then resets the equation arraylist. This will catch most calculation errors. I have not checked for text input errors or problems as of yet. Next is to try to save and retrieve the arraylist to a text file.