Index Update 204.11.10@12:37AM
In this update, stylistic changes are being put in effect. The new jar, found at http://www.bootsfaces.net/page/index.xhtml needs to be saved to the library file. This jar will allow for simplistic validation in the textbox and in the operand box. That will be seen in the next update!