Index Update 204.11.15@12:22PM
This is a facial overhaul and this update adds AJAX code to the numeric keypad. This allows for instant appearance in the Operand Display area. The Operand Display area has been changed from an inputText element to an outputText element and will require more formatting changes - ie. a box around it to make it stand out from the background, possibly larger and different font text.
The form action when the Enter key is pressed causes the del function to fire. That needs to be trapped and controlled.