loadIt.xhtml 2014.11.19@16:57EST
This is a new page that allows user to pick file for uploading. This may need to be incorporated into final calculator view.
This code was found on javatutorials.com JSF2.2 File Upload Example! This code may need to be moved to index.xhtml and be included in calculator form.