Control Update 2014.11.29@18:08
This file is all but complete. Files can be successfully upload, with validation of file and contents and integrated into the application. Files can be saved to the local machine. So far, this has only been used on same machine as web server. This may not get deployed to external web server during this course. All that is left is minor beautification.