Saver Initial Commit 2014.11.29
This file saves the current outputDisplay to a file at a certain directory in a certain drive on Windows systems. C:/TabCalc/filename.txt. Where filename is a concatenation of the date, time, and "TeAmazing_CalcTab.txt". This means unless you save the file twice at the exact same time, it will never have the same name. (Or you change your system time).
Any file saved through this class will be able to be reloaded without and concern.